Add support to async RPC handler.
A new AsyncRpcHandler type is added.

The network is different from other RPC systems, because

1. It must support adding/removing servers on the fly. The service dispatch
table is not static, and thus has to be guarded by a mutex.
2. It must simultaneously support more than one concrete types that handle
RPC. All those types are stored in the internal dispatch table, which must
use type erasure instead of generic.

There are two ways to store an RPC handler in the dispatch table
1. Arc<dyn Fn() + Send + Sync>
2. Box<dyn Factory<dyn FnOnce() + Send>>

Note 'dyn FnOnce()' cannot be clone. We chose the first one because the
second approach needs too many 'dyn's.